+++ bin/entry.sh
@@ -34,8 +34,6 @@ stop >/dev/null 2>&1 || true
 rm -f ${LOGS_DIR}/*
 rm -f ${RUN_DIR}/*
 
-export RC_CHROME="false"
-export RC_FIREFOX="false"
 # Support restart docker container for selenium 3 @aituganov
 if [ ! -f /usr/bin/geckodriver ]; then
   sudo mv /opt/geckodriver /usr/bin/geckodriver
@@ -45,11 +43,6 @@ fi
 sudo cp /capabilities3.json /capabilities.json
 sudo cp /capabilities3.json /home/seluser/capabilities.json
 sudo cp /capabilities3.json /home/seluser/caps.json
-
-# We need larger screens for Selenium IDE RC tests
-# if [ "${RC_CHROME}" = "true" ] || [ "${RC_FIREFOX}" = "true" ]; then
-#   export SCREEN_HEIGHT=$((SCREEN_HEIGHT*2))
-# fi
 
 #---------------------
 # Fix/extend ENV vars
@@ -157,14 +150,6 @@ if [ "${FIREFOX}" = "true" ]; then
   export SUPERVISOR_REQUIRED_SRV_LIST="${SUPERVISOR_REQUIRED_SRV_LIST}|selenium-node-firefox"
 fi
 
-if [ "${RC_CHROME}" = "true" ]; then
-  export SUPERVISOR_REQUIRED_SRV_LIST="${SUPERVISOR_REQUIRED_SRV_LIST}|selenium-rc-chrome"
-fi
-
-if [ "${RC_FIREFOX}" = "true" ]; then
-  export SUPERVISOR_REQUIRED_SRV_LIST="${SUPERVISOR_REQUIRED_SRV_LIST}|selenium-rc-firefox"
-fi
-
 if [ "${SELENIUM_HUB_PORT}" = "" ]; then
   echo "FATAL: SELENIUM_HUB_PORT is empty but should be a number" 1>&2
   exit 120
@@ -216,24 +201,6 @@ elif [ "${PICK_ALL_RANDOM_PORTS}" = "true" ]; then
   # User want to pick random ports but may also want to fix some others
   if [ "${SELENIUM_NODE_FF_PORT}" = "${DEFAULT_SELENIUM_NODE_FF_PORT}" ]; then
     export SELENIUM_NODE_FF_PORT=$(get_unused_port_from_range ${RANDOM_PORT_FROM} ${RANDOM_PORT_TO})
-  fi
-fi
-
-if [ "${SELENIUM_NODE_RC_CH_PORT}" = "0" ]; then
-  export SELENIUM_NODE_RC_CH_PORT=$(get_unused_port_from_range ${RANDOM_PORT_FROM} ${RANDOM_PORT_TO})
-elif [ "${PICK_ALL_RANDOM_PORTS}" = "true" ]; then
-  # User want to pick random ports but may also want to fix some others
-  if [ "${SELENIUM_NODE_RC_CH_PORT}" = "${DEFAULT_SELENIUM_NODE_RC_CH_PORT}" ]; then
-    export SELENIUM_NODE_RC_CH_PORT=$(get_unused_port_from_range ${RANDOM_PORT_FROM} ${RANDOM_PORT_TO})
-  fi
-fi
-
-if [ "${SELENIUM_NODE_RC_FF_PORT}" = "0" ]; then
-  export SELENIUM_NODE_RC_FF_PORT=$(get_unused_port_from_range ${RANDOM_PORT_FROM} ${RANDOM_PORT_TO})
-elif [ "${PICK_ALL_RANDOM_PORTS}" = "true" ]; then
-  # User want to pick random ports but may also want to fix some others
-  if [ "${SELENIUM_NODE_RC_FF_PORT}" = "${DEFAULT_SELENIUM_NODE_RC_FF_PORT}" ]; then
-    export SELENIUM_NODE_RC_FF_PORT=$(get_unused_port_from_range ${RANDOM_PORT_FROM} ${RANDOM_PORT_TO})
   fi
 fi
 
@@ -332,8 +299,6 @@ ga_track_start () {
     START_META_DATA="${START_META_DATA} SELENIUM_NODE_HOST='${SELENIUM_NODE_HOST}'"
     START_META_DATA="${START_META_DATA} SELENIUM_HUB_PARAMS='${SELENIUM_HUB_PARAMS}'"
     START_META_DATA="${START_META_DATA} SELENIUM_NODE_PARAMS='${SELENIUM_NODE_PARAMS}'"
-    START_META_DATA="${START_META_DATA} RC_CHROME='${RC_CHROME}'"
-    START_META_DATA="${START_META_DATA} RC_FIREFOX='${RC_FIREFOX}'"
     START_META_DATA="${START_META_DATA} MEM_JAVA_PERCENT='${MEM_JAVA_PERCENT}'"
     START_META_DATA="${START_META_DATA} WAIT_TIMEOUT='${WAIT_TIMEOUT}'"
     START_META_DATA="${START_META_DATA} WAIT_FOREGROUND_RETRY='${WAIT_FOREGROUND_RETRY}'"
@@ -548,10 +513,6 @@ echo "${SELENIUM_NODE_CH_PORT}" > SELENIUM_NODE_CH_PORT
 echo "${SELENIUM_NODE_CH_PORT}" > CH_PORT
 echo "${SELENIUM_NODE_FF_PORT}" > SELENIUM_NODE_FF_PORT
 echo "${SELENIUM_NODE_FF_PORT}" > FF_PORT
-echo "${SELENIUM_NODE_RC_CH_PORT}" > SELENIUM_NODE_RC_CH_PORT
-echo "${SELENIUM_NODE_RC_CH_PORT}" > RC_CH_PORT
-echo "${SELENIUM_NODE_RC_FF_PORT}" > SELENIUM_NODE_RC_FF_PORT
-echo "${SELENIUM_NODE_RC_FF_PORT}" > RC_FF_PORT
 echo "${VNC_PORT}" > VNC_PORT
 echo "${NOVNC_PORT}" > NOVNC_PORT
 echo "${SUPERVISOR_HTTP_PORT}" > SUPERVISOR_HTTP_PORT
@@ -564,8 +525,6 @@ echo "${CHROME_VERSION}" > CHROME_VERSION
 echo "${CHROME}" > CHROME
 echo "${CHROME_FLAVOR}" > CHROME_FLAVOR
 echo "${FIREFOX}" > FIREFOX
-echo "${RC_CHROME}" > RC_CHROME
-echo "${RC_FIREFOX}" > RC_FIREFOX
 echo "${WAIT_TIMEOUT}" > WAIT_TIMEOUT
 echo "${COMMON_CAPS}" > COMMON_CAPS
 echo "${SELENIUM_HUB_HOST}" > SELENIUM_HUB_HOST